fix: get total elements of table

+++ src/main/java/com/erp/shared/dto/pagination/ResponsePageDTO.java
@@ -4,6 +4,6 @@ public record ResponsePageDTO(
         Integer number,
         Integer size,
         Integer pages,
-        Integer items
+        Long items
 ) {
 }
