+++ src/main/java/com/erp/shared/dto/pagination/ResponsePaginationDTO.java
@@ -6,7 +6,7 @@ public record ResponsePaginationDTO<E>(
         ResponsePageDTO page,
         List<E> data
 ) {
-    public static <E> ResponsePaginationDTO create(Integer pageNumber, Integer pageSize, Integer totalPages, Integer totalElements, List<E> data) {
+    public static <E> ResponsePaginationDTO create(Integer pageNumber, Integer pageSize, Integer totalPages, Long totalElements, List<E> data) {
         return new ResponsePaginationDTO(
                 new ResponsePageDTO(
                         pageNumber,
